requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
click==8.3.0
croniter==6.0.0
exceptiongroup==1.3.0
fastapi==0.120.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.2.6
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.0.0
requests==2.32.3
openpyxl==3.1.3
rq==2.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
